Update __tests__/env.ts to handle initiliazation of variables
